--- a/README.md
+++ b/README.md
@@ -63,7 +63,7 @@ djin:git/dust> dust
 
 ## Performance
 
-Dust is currently about 4 times slower than du.
+Dust uses a parallel fetching implementation that greatly improves performance for directory trees with reasonable amount of files (read more than 20) compared to du. This can be as much as 7x faster than du on a clean cache.
 
 ## Alternatives