@AdminXVII 2019-11-26
The problem was that JWalk does not take hidden directories in account by default.

On a side note, do you want me to add a CLI flag to ignore hidden directories? That could be a nice feature.