@bootandy 2019-12-03
You were correct about the hidden directories. Nice.

I have been running your branch vs the existing branch locally and I see very little difference in execution speed. I certainly can't get anywhere near the speed of du so I'd like to know how you benchmarked that.

Update: I have been able to get slightly superior performance to the existing dust by limiting the number of threads created to approximately match my cpu cores. I think the thing to do might be to incorporate something like: https://crates.io/crates/num_cpus to ensure we don't spawn too many threads.

I don't think we need a flag to ignore hidden directories, (I personally cant see the usecase, but if several people ask I'll add it). However, I would like an optional flag to control the aforementioned number of CPU cores we use (Personally I don't like it when a tool always grabs all your cpu cores). 

I do like the commit:  `Use more rusty patterns and preallocate enough space ` I think that improves things. 

I'll wait a week to see if you want to add anything and if not I'll merge and extend your work myself.

Cheers,