@AdminXVII 2019-12-03
The performance surely depends on your hardware, I have a SSD with a 8 core intel. Maybe HDD and/or a lower number of cores is the reason you see the difference in speed.

For the benchmark used, I did
```
[parallel-walk][~/dev/dust]$ cargo run --release # dont benchmark cargo
[parallel-walk][~/dev/dust]$ time ./target/release/dust ..
  57G ─┬ ..
  15G  ├─┬ redox
 5.7G  │ ├─┬ cookbook
 4.9G  │ │ └── recipes
 3.0G  │ └─┬ .git
 3.0G  │   └─┬ modules
 2.9G  │     └── rust
 6.1G  ├─┬ shellac-server
 5.4G  │ └─┬ target
 3.0G  │   └── debug
 6.1G  ├─┬ robotique
 6.1G  │ └─┬ target
 4.2G  │   └─┬ debug
 3.2G  │     └── deps
 5.8G  ├─┬ iced
 5.4G  │ └─┬ target
 4.7G  │   └─┬ debug
 2.7G  │     └── deps
 3.3G  ├─┬ planet-gen
 3.3G  │ └── target
 2.6G  └── ion
real    5.535719165s
[parallel-walk][~/dev/dust]$ time du -sh ..
58G	..
real    21.578478589s
```

jwalk uses rayon, which states:
"By default, Rayon uses the same number of threads as the number of CPUs available. Note that on systems with hyperthreading enabled this equals the number of logical cores and not the physical ones."
So theorically, the number of threads should already match the number of cores.

Would RAYON_NUM_THREADS be enough for you? If not, I'll try to update the MR before this weekend.